fixed type statement in doc

+++ starfish/pipeline/filter/gaussian_high_pass.py
@@ -55,7 +55,7 @@ class GaussianHighPass(FilterAlgorithmBase):
 
         Parameters
         ----------
-        image : numpy.ndarray[np.uint32]
+        image : numpy.ndarray[np.uint16]
             2-d or 3-d image data
         sigma : Union[Number, Tuple[Number]]
             Standard deviation of gaussian kernel
